pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tg-bot-chat"
version = "0.1.0"
description = "Telegram bot with chat context management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot[all]>=21.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.hatch.build.targets.wheel]
packages = ["src/tg_bot_chat"] 
